ai medicine and disease prediction/ml_service/requirements.txt
# ─── Core Flask ───────────────────────────────────────────────────────────────
Flask==2.3.3
flask-cors==4.0.0

# ─── ML / Data Science ────────────────────────────────────────────────────────
numpy==1.25.2
pandas==2.1.0
scikit-learn==1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# ─── Image Processing / OCR ───────────────────────────────────────────────────
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0

# ─── AI / NLP ─────────────────────────────────────────────────────────────────
google-generativeai>=0.3.1
g4f>=0.4.3.4
langdetect>=1.0.9

# ─── Speech (optional) ────────────────────────────────────────────────────────
SpeechRecognition>=3.10.0
gTTS>=2.4.0

# ─── PDF Report Generation ────────────────────────────────────────────────────
fpdf2>=2.7.6

# ─── Utilities ────────────────────────────────────────────────────────────────
aiohttp>=3.9.0
python-dotenv>=1.0.0
